EDA/online-news-popularity/pyproject.toml
# ---------------------------------------------------------------------------
# Project metadata (PEP 621)
# ---------------------------------------------------------------------------
[project]
name            = "online_news_popularity"
version         = "0.1.0"               # keep in sync with __init__.__version__
description     = "Modular cleaning pipeline for the Online News Popularity dataset"
readme          = "README.md"
requires-python = ">=3.9"
license         = {text = "MIT"}
authors = [
  {name = "Your Team 41", email = "team41@example.com"}
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

# ---------------------------------------------------------------------------
# Dependencies
# ---------------------------------------------------------------------------
dependencies = [
  "pandas>=1.5",
  "numpy>=1.22",
  "pyyaml>=6.0",
  "fsspec>=2023.4.0",     # remote I/O abstraction (optional back-ends via extras)
]

[project.optional-dependencies]
viz      = ["matplotlib>=3.7", "seaborn>=0.13"]
hydra    = ["hydra-core>=1.3", "omegaconf>=2.3"]
dev      = [
  "pytest>=7.4",
  "black>=23.9",
  "ruff>=0.1.0",
  "mypy>=1.5",
  "build>=0.10",          # python -m build
  "twine>=4.0",
]

# Convenience console-script (so users can type `onp-clean …`)
[project.scripts]
onp-clean = "online_news_popularity.pipeline.clean_dataset:main"

# ---------------------------------------------------------------------------
# Build backend
# ---------------------------------------------------------------------------
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

# ---------------------------------------------------------------------------
# Tooling configs (Black, Ruff, Mypy) – keep it minimal
# ---------------------------------------------------------------------------
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E203", "E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
